+++ keyboards/ergohaven/ergohaven_main.c
@@ -1,6 +1,6 @@
 #include "ergohaven.h"
 #include "src/eh_settings.h"
-#include "ergohaven_ruen.h"
+#include "src/eh_ruen.h"
 #include "src/oled/eh_oled.h"
 #include "ergohaven_rgb.h"
 #include "src/display/eh_display.h"
