+++ keyboards/ergohaven/hpd/keymaps/v2_lcd_ball/rules.mk
@@ -24,6 +24,6 @@ SRC += keyboards/ergohaven/src/display/fonts/eh_font_montserrat_28.c
 
 
 SRC += keyboards/ergohaven/ergohaven_main.c
-SRC += keyboards/ergohaven/ergohaven_ruen.c
+SRC += keyboards/ergohaven/src/eh_ruen.c
 SRC += keyboards/ergohaven/hid.c
 SRC += keyboards/ergohaven/ergohaven_pointing.c
