+++ keyboards/ergohaven/hpd/rev1/rules.mk
@@ -30,6 +30,6 @@ UNICODE_ENABLE = yes
 
 SRC += keyboards/ergohaven/ergohaven_main.c
 SRC += keyboards/ergohaven/src/eh_settings.c
-SRC += keyboards/ergohaven/ergohaven_ruen.c
+SRC += keyboards/ergohaven/src/eh_ruen.c
 SRC += keyboards/ergohaven/hid.c
 SRC += keyboards/ergohaven/ergohaven_pointing.c
