+++ keyboards/ergohaven/sm30/rules.mk
@@ -37,7 +37,7 @@ UNICODE_COMMON = yes
 UNICODE_ENABLE = yes
 
 SRC += keyboards/ergohaven/ergohaven_main.c
-SRC += keyboards/ergohaven/ergohaven_ruen.c
+SRC += keyboards/ergohaven/src/eh_ruen.c
 SRC += keyboards/ergohaven/hid.c
 SRC += keyboards/ergohaven/ergohaven_pointing.c
 SRC += keyboards/ergohaven/src/eh_settings.c
