+++ keyboards/ergohaven/src/display/eh_keycode_str.c
@@ -3,7 +3,7 @@
 #include "keycodes.h"
 #include "ergohaven.h"
 #include "eh_symbols.h"
-#include "ergohaven_ruen.h"
+#include "src/eh_ruen.h"
 #include "ergohaven_pointing.h"
 
 const char *basic_keycode_to_str(uint16_t keycode) {
